move all types and interfaces in their own files
