+++ src/ioc/token.ts
@@ -1,15 +1,8 @@
+import type { MaybeToken, Token } from "./types";
+
 export function token<T>(name: string) {
     return {type: Symbol(name)} as Token<T>;
 }
-
-declare const typeMarker: unique symbol;
-
-export interface Token<T> {
-    type: symbol;
-    [typeMarker]: T;
-}
-
-export type MaybeToken<T = unknown> = Token<T> | symbol;
 
 function isToken<T>(token: MaybeToken<T>): token is Token<T> {
     return typeof token != "symbol";
