+++ src/ioc/wire.ts
@@ -1,6 +1,6 @@
 import {Container} from "./container";
 import {define} from "./define";
-import {MaybeToken} from "./token";
+import type {MaybeToken} from "./types";
 
 export function createWire(container: Container) {
     return <Value, Target extends {[key in Prop]: Value}, Prop extends string>(
